pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "fotoerp-backend"
version = "0.1.0"
description = "FotoDerp Backend — Bildanalyse, Metadaten, KI-Pipeline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "redis>=5.2",
    "pillow>=11.0",
    "exifread>=3.0",
    "rawpy>=0.21",
    "httpx>=0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "rich>=14.0",
    "click>=8.1",
    "numpy>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
    "mypy>=1.13",
]

[project.scripts]
fotoerp-backend = "fotoerp_backend.cli:main"

[tool.setuptools.packages.find]
include = ["fotoerp_backend*"]
exclude = ["tests*", "scripts*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A"]

[tool.mypy]
python_version = "3.11"
strict = true
